refactor: useSiteSettingsStore разбит на отдельные composables
